Initial write support (#122)
* Zigzag and varint encoders

* Short repeat and direct RLEv2 writers

* Minor refactoring

* Signed msb encoder

* Refacftor RLE delta+patched base to be more functional

* Minor refactoring

* Byte RLE writer

* Remove unused error types

* Initial version of ORC writer, supporting only float

* Int8 array write support

* Integer RLEv2 Delta writing

* Minor optimization

* Abstract bits_used functionality to common NInt function

* Remove overcomplicated AbsVarint code, replace with i64/u64 in delta encoding

* Minor fixes

* Initial RLEv2 encoder base

* Remove u64 impl of NInt in favour of generic to determine sign

* Simplify getting RLE reader

* Fix percentile bit calculation encoding/decoding

* Patched base writing

* Support writing int arrays

* Multi stripe write test case

* Reduce duplication for primitive ColumnStripeEncoders

* Introduce EstimateMemory trait to standardize

* Comment

* Remove need for seek from writer + minor PR comments

* Rename s_type to kind

* Deduplicate get_closest_fixed_bits

* Fix comments

* Switch arrow writer tests to be in-memory instead of writing to disk

* Fix writing arrays with nulls

* Add license to new files

+++ src/reader/decode/timestamp.rs
@@ -27,7 +27,7 @@ const NANOSECONDS_IN_SECOND: i64 = 1_000_000_000;
 pub struct TimestampIterator<T: ArrowTimestampType, Item: TryFrom<i128>> {
     base_from_epoch: i64,
     data: Box<dyn Iterator<Item = Result<i64>> + Send>,
-    secondary: Box<dyn Iterator<Item = Result<u64>> + Send>,
+    secondary: Box<dyn Iterator<Item = Result<i64>> + Send>,
     _marker: PhantomData<(T, Item)>,
 }
 
@@ -35,7 +35,7 @@ impl<T: ArrowTimestampType, Item: TryFrom<i128>> TimestampIterator<T, Item> {
     pub fn new(
         base_from_epoch: i64,
         data: Box<dyn Iterator<Item = Result<i64>> + Send>,
-        secondary: Box<dyn Iterator<Item = Result<u64>> + Send>,
+        secondary: Box<dyn Iterator<Item = Result<i64>> + Send>,
     ) -> Self {
         Self {
             base_from_epoch,
@@ -61,10 +61,11 @@ impl<T: ArrowTimestampType, Item: TryFrom<i128>> Iterator for TimestampIterator<
 fn decode_timestamp<T: ArrowTimestampType, Ret: TryFrom<i128>>(
     base: i64,
     seconds_since_orc_base: Result<i64>,
-    nanoseconds: Result<u64>,
+    nanoseconds: Result<i64>,
 ) -> Result<Option<Ret>> {
     let data = seconds_since_orc_base?;
-    let mut nanoseconds = nanoseconds?;
+    // TODO
+    let mut nanoseconds = nanoseconds? as u64;
     // Last 3 bits indicate how many trailing zeros were truncated
     let zeros = nanoseconds & 0x7;
     nanoseconds >>= 3;
